[TASK] Use additional secret for apache delivery token
The access token is generated as a HMAC from the user
ids. As this would be valid for future requests
an additional token is introduced that will be cached
in configuration if the page is cached or
generated again on every rendering when the page is
not cached.

This can generate many files and folders and requires
the garbage collect feature to be implemented to
mitigate that.