WIP|FEATURE: Use download HTML attribute if available
This is a proof of concept change to allow HTML5 download attribute to
be respected. This allows to not force download of files in general, but
to turn it on when the attribute is set.

Technically this change affects way to many parts of the software and
introduces a global variable.

Proposed changes:

1. Hashing should be moved to a single component in order to reduce
   places to touch and places containing hashing logic.
2. Parsing of HTML should be improved. Always use HTML Tag parsing
   through TYPO3 API and allow a whitelist of HTML tags to pre
   processed. Right now the parsing is quite unstable.

Resolves: #45