[TASK] Check for correct value in CheckConfiguration
If something goes wrong when evaluating the path for a directory
(e.g. in the case of a dangling symbolic link) a value of false
is returned for the path. This would throw an exception if
proceeded with this value.

This patch checks if a value is returned before continuing and
prevents the exception.

Resolves: #130