[TASK] Optimization in CheckConfiguration (max 20)
In CheckConfiguration (which is called when the Extension
Configuration dialog is opened), the filesystem is scanned
and a collection of files assembled which are not protected.
This only collects 20 files in any case, but the outer loop
continues - even if the 20 files limit is already reached.

This patch changes the behaviour, so that rendering is
started as soon as the 20 files limit is reached.

Related: #133