PHP 8.1 compatibility (#137)
* [TASK] TYPO3 v11 support

Resolves: #126
Resolves: #119

* [TASK] Optimization in CheckConfiguration (max 20)

In CheckConfiguration (which is called when the Extension
Configuration dialog is opened), the filesystem is scanned
and a collection of files assembled which are not protected.
This only collects 20 files in any case, but the outer loop
continues - even if the 20 files limit is already reached.

This patch changes the behaviour, so that rendering is
started as soon as the 20 files limit is reached.

Related: #133

* [TASK] Check for correct value in CheckConfiguration

If something goes wrong when evaluating the path for a directory
(e.g. in the case of a dangling symbolic link) a value of false
is returned for the path. This would throw an exception if
proceeded with this value.

This patch checks if a value is returned before continuing and
prevents the exception.

Resolves: #130

* [TASK] allow PHP 8.1 in ext_emconf.php, add php dep to composer.json

* [BUGFIX] $GLOBALS['TSFE'] not available in backend

PHP Warning: Undefined global variable $TSFE in /typo3conf/ext/secure_downloads/Classes/Factory/SecureLinkFactory.php line 61, 80

* [BUGFIX] Array to string conversion

Array to string conversion in //secure_downloads/Classes/Domain/Transfer/Token/AbstractToken.php line 133

Co-authored-by: Sybille Peters <sypets@gmx.de>