Release 7.0.2 prepare (#256)
* [DOCS] Prepare release of version 7.0.2

This commit adds the changelog documentation for the 7.0.2 release,
which includes bugfixes and maintenance updates. The changelog lists
all changes since version 7.0.1, including documentation improvements,
FAL driver registration fix, and prevention of undefined array access.

* [DOCS] Update changelog for version 7.0.2 (#253)

This commit finalizes the changelog documentation for the 7.0.2 release
by listing all the fixes and improvements included in this version:
- Fix typo in documentation
- Register FAL driver correctly
- Prevent undefined array access
- Remove accidental citations

Additionally, the changelog index has been updated to include the 7-0-2
changelog page in the documentation navigation.